Use UTF-8; fix some C4244 warnings

+++ include/triton/Dialect/Triton/IR/Utility.h
@@ -20,7 +20,7 @@ SmallVector<T> convertType(const VecU &in) {
 }
 
 template <typename Int> Int product(llvm::ArrayRef<Int> arr) {
-  return std::accumulate(arr.begin(), arr.end(), 1, std::multiplies{});
+  return std::accumulate(arr.begin(), arr.end(), (Int)1, std::multiplies{});
 }
 template <typename VecT> auto product(const VecT &vec) {
   return product(llvm::ArrayRef(vec));
